fix siamese bug

+++ paddlehub/finetune/task/pairwise_task.py
@@ -516,7 +516,7 @@ class PairwiseTask(BaseTask):
         results = []
         for batch_state in run_states:
             batch_result = batch_state.run_results
-            results += batch_result[0].tolist()
+            results += batch_result[0].astype(int).tolist()
             # for sample_infer in batch_result:
             #     results.append(sample_infer.tolist()[0])
         return results
